update project and clean code accounts
